\name{sparsemixmat-package}
\alias{sparsemixmat-package}
\alias{sparsemixmat}
\docType{package}
\title{
\packageTitle{sparsemixmat}
}
\description{
\packageDescription{sparsemixmat}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sparsemixmat}
\packageIndices{sparsemixmat}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sparsemixmat}

Maintainer: \packageMaintainer{sparsemixmat}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
