% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_matrix_data}
\alias{scale_matrix_data}
\title{Scale and centering of three-way objects}
\usage{
scale_matrix_data(X)
}
\arguments{
\item{X}{p \times q \times N three-way data}
}
\value{

}
\description{
Scale and centering of three-way objects
}
